+++ CHANGELOG.md
@@ -20,6 +20,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Deprecated `ui.always-allow-large-revsets` setting and `all:` revset modifier
   have been removed.
 
+* `<name>@<remote>` revset symbols can also be resolved to remote tags. Tags are
+  prioritized ahead of bookmarks.
+
 * Legacy placeholder support used for unset `user.name` or `user.email` has been
   removed. Commits containing these values will now be pushed with `jj git push`
   without producing an error.